处理时间以及下划线驼峰等问题

+++ showExcel/src/main/java/com/showexcel/model/PowerCashStatistics.java
@@ -4,6 +4,6 @@ public class PowerCashStatistics {
 
     private CashStatistics cashStatistics;
 
-    private Double holidayTemporaryReceipt;
+    private BigDecimal holidayTemporaryReceipt;
 
 }
